// customer-form/form/types.ts
export interface FormData {
  firstName: string;
  lastName: string;
  email: string;
  whatsapp: string;
  brandName: string;
  website: string;
  whatSell: string;
  whatSource: string;
  orderVolume: string;
  expectedOrderSize: string;
  country: string;
  experience: string;
  references: File[];
  budget: string;
  timeline: string;
}